+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 新增 `checkbox` 组件
 - 新增 `rate` 组件
 - `f-radio-group` 组件废除 `circle` 配置项
+- 修复 `f-radio` 组件 `border` 状态禁用样式细节
 
 ## 0.7.0-alpha.3 (2022-10-06)